requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
httpx==0.24.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
Jinja2==3.1.2
itsdangerous==2.1.2
pydantic==1.10.12

